Add Loader service functionality to tt::app:: namespace and make loader pubsub usable.
